chore(steppers): reorganize directory structure
