+++ packages/accordions/examples/basic.md
@@ -1,4 +1,4 @@
-### Steppers
+### Stepper
 
 Stepper implementations can use [ARIA live regions](https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/ARIA_Live_Regions)
 to communicate what step the user is on. For a horizontal `Stepper`, render step content outside of
