+++ packages/accordions/src/styled/stepper/StyledContent.spec.tsx
