+++ packages/accordions/src/styled/stepper/StyledContent.ts
