+++ packages/accordions/src/styled/stepper/StyledIcon.spec.tsx
