+++ packages/accordions/src/styled/stepper/StyledIcon.ts
