+++ packages/accordions/src/styled/stepper/StyledInnerContent.tsx
