+++ packages/accordions/src/styled/stepper/StyledLabel.spec.tsx
