+++ packages/accordions/src/styled/stepper/StyledLabel.ts
