+++ packages/accordions/src/styled/stepper/StyledLabelText.spec.tsx
