+++ packages/accordions/src/styled/stepper/StyledLabelText.ts
