+++ packages/accordions/src/styled/stepper/StyledLine.ts
