+++ packages/accordions/src/styled/stepper/StyledStep.spec.tsx
