+++ packages/accordions/src/styled/stepper/StyledStep.ts
