+++ packages/accordions/src/styled/stepper/StyledStepper.spec.tsx
