+++ packages/accordions/src/styled/stepper/StyledStepper.ts
